+++ report/src/main/webapp/app/app.config.js
@@ -47,6 +47,7 @@ require.config({
     'metadataService': 'services/metadata.service',
     'metadataAccessService': 'services/metadataAccess.service',
     'notesService': 'services/notes.service',
+    'historyService': 'services/history.service',
     'suiteInfoService': 'services/suiteInfo.service',
     'patternsService': 'services/patterns.service',
     'userSettingsService': 'services/userSettings.service',
@@ -73,7 +74,8 @@ require.config({
     'includedCommentPopoverDirective': 'layout/main/url/includedCommentPopover.directive',
     //modals
     'unsavedChangesModalController': 'layout/modal/unsavedChanges/unsavedChangesModal.controller',
-    'noteModalController': 'layout/modal/note/noteModal.controller'
+    'noteModalController': 'layout/modal/note/noteModal.controller',
+    'historyModalController': 'layout/modal/history/historyModal.controller'
   },
   shim: {
     jquery: {