+++ rest-endpoint/src/main/java/com/cognifide/aet/rest/BasicDataServlet.java
@@ -86,6 +86,19 @@ abstract class BasicDataServlet extends HttpServlet {
     return ValidatorProvider.getValidator().validateValue(Suite.class, "name", suiteName).isEmpty();
   }
 
+  boolean isValidVersion(String version) {
+    boolean valid = false;
+    if (version != null) {
+      try {
+        Integer.parseInt(version);
+        valid = true;
+      } catch (NumberFormatException ex) {
+        // Ok, continue
+      }
+    }
+    return valid;
+  }
+
   boolean isValidCorrelationId(String correlationId) {
     return ValidatorProvider.getValidator()
         .validateValue(Suite.class, "correlationId", correlationId).isEmpty();