+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to AET will be documented in this file.
 ## Unreleased
 **List of changes that are finished but not yet released in any final version.**
 
+- [PR-328](https://github.com/Cognifide/aet/pull/328) Added suite's history
 - [PR-303](https://github.com/Cognifide/aet/pull/303) Added `exclude-elements` parameter to [ScreenCollector](https://github.com/Cognifide/aet/wiki/ScreenCollector)
 - [PR-327](https://github.com/Cognifide/aet/pull/327) Default web driver changed from Firefox to Chrome
 - [PR-294](https://github.com/Cognifide/aet/pull/294) Added support for full page screenshot in chrome
@@ -25,6 +26,9 @@ All notable changes to AET will be documented in this file.
 - [PR-291](https://github.com/Cognifide/aet/pull/291) Updated nu.validator version from 15.6.29 to 17.11.1
 - [PR-298](https://github.com/Cognifide/aet/pull/298) Filtering accessibility errors by markup CSS ([#214](https://github.com/Cognifide/aet/issues/214))
 - [PR-296](https://github.com/Cognifide/aet/pull/296) Status Code Comparator now checks range 400-600 by default, parameters validation added
+- [PR-325](https://github.com/Cognifide/aet/pull/325) Support env variables in report manager
+- [PR-331](https://github.com/Cognifide/aet/pull/325) Support env variables in MongoDB config
+- [PR-338](https://github.com/Cognifide/aet/pull/338) Enable creating new dbs by default
 
 ## Version 2.1.6